Test: (BRD-132) 게시글 카테고리 생성 API 명세 문서화(RestDocs)
