+++ board-system-board/board-web-adapter/src/test/kotlin/com/ttasjwi/board/system/board/test/base/BoardRestDocsTest.kt
@@ -6,4 +6,3 @@ import org.springframework.context.annotation.Import
 
 @Import(BoardRestDocsTestConfig::class)
 abstract class BoardRestDocsTest : RestDocsTest()
-
